🚨 Fix: Footer width 수정 외 2건
- 각 아이콘 link 속성 추가 및 할당
- 명상 페이지 아이콘 변경
- Icon 크기 증가
