+++ src/components/Footer/index.ts
@@ -1,0 +1,3 @@
+import { Footer } from '@components/Footer/Footer';
+
+export default Footer;
